Update to latest comments
Some real hard git-black-magic had to happen on my end to set this up (it shouldn't be an issue going forward), so there may be some weirdness in this commit --- we'll find out in a second!

Otherwise, this is just a bunch of QOL/optimization style changes with a focus on best practices. I also updated a few fields that I felt should have been serialized.